requirements.txt
openai==0.27.3
python-dotenv==1.0.0